Test improvements and some bug fixes
Fixed a rare bug with the growing algorithm.
Fixed a bug with compose (lowCompose and highCompose where not pushed to the work stack)
Added theory tests for compose and support based on syntax trees.
Reworked tests to get an even spread to each type of operations (previously unary, binary and ternary tests had n, n^2 and n^3 invocations, resp.)